move infra to deploy
