+++ deploy/sql/dummy/school.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const sql = require('mssql')
-const config = require('../config')
+const config = require('../../../infra/config')
 
 const schoolCount = 20000
