+++ deploy/sql/readme.md
@@ -1,0 +1,10 @@
+
+# SQL utilities
+
+## Database migrator
+
+- duplicate of `admin/data/sql` that allows targeting custom database and server via arguments
+- seeding is broken due to reliance on sql service (work in progress)
+- example baseline schema and data scripts are under `baseline` folder
+
+
