+++ deploy/sql/shell/readme.md
@@ -1,0 +1,3 @@
+cannot find any reference to these scripts in VSO
+
+will be deleted if found to be unused
