--- infra/readme.md
@@ -1,6 +1,0 @@
-
-# MTC Infrastructure
-
-scripts and utilities to deploy / manage components for local development and cloud hosting.
-
-Azure CLI is preferred over ARM templates due to its terseness and flexibility.
